fix : hotel type error
Former-commit-id: a465c93771c55bf12bd7972ff03d4724724b280d
Former-commit-id: c261e10b1d216df9e024cc7492f16b3562c60624

+++ backend/src/main/java/com/ssafy/i5i/hotelAPI/domain/food/controller/FoodController.java
@@ -21,13 +21,13 @@ public class FoodController {
 	private final FoodServiceImpl foodServiceImpl;
 
 	@GetMapping("/{name}/{distance}/{sorted}")
-	public List<FoodResponseDto> getFoodFromTravle(@PathVariable("name")String name, @PathVariable("distance")Double distance, @PathVariable("sorted")String sorted ){
+	public List<FoodResponseDto> getFoodFromTravle(@PathVariable("name")String name, @PathVariable("distance")Integer distance, @PathVariable("sorted")String sorted ){
 		AttractionNameRequestDto attractionNameRequestDto = new AttractionNameRequestDto(name, distance, sorted);
 		return foodServiceImpl.getFoodFromTravle(attractionNameRequestDto);
 	}
 
 	@GetMapping("/{longtitude}/{latitude}/{distance}/{sorted}")
-	public List<FoodResponseDto> getFoodFromLngLat(@PathVariable("longtitude")Double longitude, @PathVariable("latitude")Double latitude, @PathVariable("distance")Long distance, @PathVariable("sorted")String sorted){
+	public List<FoodResponseDto> getFoodFromLngLat(@PathVariable("longtitude")Double longitude, @PathVariable("latitude")Double latitude, @PathVariable("distance")int distance, @PathVariable("sorted")String sorted){
 		AttractionCoordinateRequestDto attractionCoordinateRequestDto = new AttractionCoordinateRequestDto(latitude, longitude, distance, sorted);
 		return foodServiceImpl.getFoodFromLngLat(attractionCoordinateRequestDto);
 	}
